[modbus] Update ValueError
Я перенёс ValueError из Sensor в Value::ValueDirect::Read
